Rename `monument::{graph::BuildError -> Error}`
Also moves `Graph::new` into the `graph::build` module.
