+++ monument/lib/src/lib.rs
@@ -19,6 +19,7 @@ use utils::{
 };
 
 use std::{
+    fmt::{Display, Formatter},
     hash::Hash,
     ops::Range,
     sync::{
@@ -194,6 +195,114 @@ impl Default for Config {
         }
     }
 }
+
+////////////
+// ERRORS //
+////////////
+
+/// The different ways that graph building can fail
+#[derive(Debug)]
+pub enum Error {
+    /// The given maximum graph size limit was reached
+    SizeLimit(usize),
+    /// Different start/end rows were specified in a multi-part
+    DifferentStartEndRowInMultipart,
+    /// The same [`Chunk`](graph::Chunk) could start at two different strokes, and some
+    /// [`MusicType`](crate::music::MusicType) relies on that
+    InconsistentStroke,
+    /// Some [`Call`](crate::Call) doesn't have enough calling positions to cover the [`Stage`]
+    WrongCallingPositionsLength {
+        call_name: String,
+        calling_position_len: usize,
+        stage: Stage,
+    },
+    /// Some [`Call`](crate::Call) refers to a lead location that doesn't exist
+    UndefinedLeadLocation { call_name: String, label: String },
+    /// [`Query`] didn't define any [`Method`]s
+    NoMethods,
+    /// Two [`Method`]s use the same shorthand
+    DuplicateShorthand {
+        shorthand: String,
+        title1: String,
+        title2: String,
+    },
+    NoCourseHeadInPart {
+        mask_in_first_part: Mask,
+        part_head: RowBuf,
+        mask_in_other_part: Mask,
+    },
+}
+
+pub type Result<T> = std::result::Result<T, Error>;
+
+impl Display for Error {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Error::SizeLimit(limit) => write!(
+                f,
+                "Graph size limit of {} chunks reached.  You can set it \
+higher with `--graph-size-limit <n>`.",
+                limit
+            ),
+            Error::DifferentStartEndRowInMultipart => {
+                write!(f, "Start/end rows must be the same for multipart comps")
+            }
+            Error::InconsistentStroke => write!(
+                f,
+                "The same chunk of ringing can be at multiple strokes, probably \
+because you're using a method with odd-length leads"
+            ),
+            Error::NoMethods => write!(f, "Can't have a composition with no methods"),
+            Error::WrongCallingPositionsLength {
+                call_name,
+                calling_position_len,
+                stage,
+            } => write!(
+                f,
+                "Call {:?} only specifies {} calling positions, but the stage has {} bells",
+                call_name,
+                calling_position_len,
+                stage.num_bells()
+            ),
+            Error::DuplicateShorthand {
+                shorthand,
+                title1,
+                title2,
+            } => write!(
+                f,
+                "Methods {:?} and {:?} share a shorthand ({})",
+                title1, title2, shorthand
+            ),
+            Error::UndefinedLeadLocation { call_name, label } => write!(
+                f,
+                "Call {:?} refers to a lead location {:?}, which doesn't exist",
+                call_name, label
+            ), // TODO: Suggest one that does exist
+            Error::NoCourseHeadInPart {
+                mask_in_first_part,
+                part_head,
+                mask_in_other_part,
+            } => {
+                writeln!(
+                    f,
+                    "course head `{}` becomes `{}` in the part starting `{}`, which isn't in `course_heads`.",
+                    mask_in_first_part, mask_in_other_part, part_head
+                )?;
+                write!(
+                    f,
+                    "   help: consider adding `{}` to `course_heads`",
+                    mask_in_other_part
+                )
+            }
+        }
+    }
+}
+
+impl std::error::Error for Error {}
+
+//////////////////
+// COMPOSITIONS //
+//////////////////
 
 /// A `Comp`osition generated by Monument.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
@@ -398,7 +507,7 @@ impl std::fmt::Display for DisplayComp<'_> {
 ////////////
 
 impl Query {
-    pub fn unoptimised_graph(&self, config: &Config) -> Result<Graph, graph::BuildError> {
+    pub fn unoptimised_graph(&self, config: &Config) -> self::Result<Graph> {
         graph::Graph::new(self, config)
     }
